Get the getTable examples to run again
The examples of fetching data via getTable and getTableWith in the
code were failing to run due to a parsing error.  It looks like the
JSON schema from Quandl has changed in the last couple of years.

The problem was with the "private" field, which is no longer part of
the Quandl schema.  There is a "premium" field in the returnd data
which I'm guessing serves the same purpose, but since it wasn't being
used anyway I decided to just get rid of it.